update unify
Update file structure, remove class re-implementation, remove redundant code
